+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Backbone.Handlebars
 
 Version 1.0.0 (unreleased)
 
+* ```{{views}}``` can render Backbone.Collections (or any object having a map method) *Radoslav Stankov*
 * Added ```{{views viewName modelsList}}``` helper *Radoslav Stankov*
 * Added ```Backbone.Handlebars``` namespace for the extention *Radoslav Stankov*
 * Added ```{{view viewName}}``` helper *Radoslav Stankov*